requirements.txt
numpy>=1.19.5
tensorflow>=2.8.0
opencv-python>=4.5.3
matplotlib>=3.4.3
Pillow>=8.3.1
scikit-image>=0.18.3
flask>=2.0.1
werkzeug>=2.0.1
scikit-learn>=0.24.2
tensorflow-addons>=0.16.1